pyproject.toml
[project]
name = "agent-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.14.1",
    "pydantic>=2.11.9",
]
